arm/dt: Create new ro.carrier for Japan
Because of Japan govern regulations we need a special NFC
config for Japan. In order to separate the Japan software
the others we are creating a new ro.carrier retjp.
And also adding retjp in the source code to get same as
retapac settings.

Change-Id: Iacf63d5cb4e4db5d24b209f7cc48165c6de50e97
Reviewed-on: https://gerrit.mot.com/1627296
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Tarciano Facco Pegoraro <tarciano@motorola.com>
Reviewed-by: Regina Mitsue Azuma <reginaazuma@motorola.com>
Reviewed-by: Ling Jin <lingjin@motorola.com>
Submit-Approved: Jira Key